Simplified SimulationRunner

+++ src/main/java/catworks/phenomena/ProbabilityMatrix.java
@@ -4,9 +4,9 @@ import java.lang.System;
 
 public class ProbabilityMatrix {
     private int[] sizes; // this contains the sizes for each network in the IDN
-    private float[][] probs; // this contains the probability of spread for each "section" of the IDN
+    private double[][] probs; // this contains the probability of spread for each "section" of the IDN
 
-    public ProbabilityMatrix(int[] sizes, float[][] probs) {
+    public ProbabilityMatrix(int[] sizes, double[][] probs) {
         this.sizes = sizes;
         this.probs = probs;
     }
@@ -15,7 +15,7 @@ public class ProbabilityMatrix {
         return sizes;
     }
 
-    public float[][] getProbs() {
+    public double[][] getProbs() {
         return probs;
     }
 }
